onboarding_ai_assistant/requirements.txt
streamlit>=1.32.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
openai>=1.3.5
